Мелкие доработки:
Изменен тип int на long везде для количества мёда
Доработан и протестирован ServiceLocator
Доработан LongOperationSimulator

+++ VS Solution/Apiary/Apiary.Client/ViewModels/Design/BeehiveVmDesignMode.cs
@@ -74,7 +74,7 @@
         /// <summary>
         /// Количество мёда.
         /// </summary>
-        public int HoneyCount { get; private set; } = 500;
+        public long HoneyCount { get; private set; } = 500;
 
         /// <summary>
         /// Количество пчёл-маток.
